Add Android support for accessibility actions.
