+++ ReactAndroid/src/main/java/com/facebook/react/uimanager/BaseViewManager.java
@@ -14,11 +14,14 @@ import java.util.HashMap;
 
 import com.facebook.react.R;
 import com.facebook.react.bridge.ReadableArray;
-import com.facebook.react.uimanager.AccessibilityDelegateUtil.AccessibilityRole;
+import com.facebook.react.common.MapBuilder;
+import com.facebook.react.uimanager.ReactAccessibilityDelegate;
+import com.facebook.react.uimanager.ReactAccessibilityDelegate.AccessibilityRole;
 import com.facebook.react.uimanager.annotations.ReactProp;
 import com.facebook.react.uimanager.util.ReactFindViewUtil;
 
 import javax.annotation.Nonnull;
+import java.util.Map;
 import javax.annotation.Nullable;
 
 /**
@@ -38,6 +41,7 @@ public abstract class BaseViewManager<T extends View, C extends LayoutShadowNode
   private static final String PROP_ACCESSIBILITY_LIVE_REGION = "accessibilityLiveRegion";
   private static final String PROP_ACCESSIBILITY_ROLE = "accessibilityRole";
   private static final String PROP_ACCESSIBILITY_STATES = "accessibilityStates";
+  private static final String PROP_ACCESSIBILITY_ACTIONS = "accessibilityActions";
   private static final String PROP_IMPORTANT_FOR_ACCESSIBILITY = "importantForAccessibility";
 
   // DEPRECATED
@@ -189,6 +193,15 @@ public abstract class BaseViewManager<T extends View, C extends LayoutShadowNode
     }
   }
 
+  @ReactProp(name = PROP_ACCESSIBILITY_ACTIONS)
+  public void setAccessibilityActions(T view, ReadableArray accessibilityActions) {
+    if (accessibilityActions == null) {
+      return;
+    }
+
+    view.setTag(R.id.accessibility_actions, accessibilityActions);
+  }
+
   @ReactProp(name = PROP_IMPORTANT_FOR_ACCESSIBILITY)
   public void setImportantForAccessibility(@Nonnull T view, @Nullable String importantForAccessibility) {
     if (importantForAccessibility == null || importantForAccessibility.equals("auto")) {
@@ -291,12 +304,19 @@ public abstract class BaseViewManager<T extends View, C extends LayoutShadowNode
   }
 
   private void updateViewAccessibility(@Nonnull T view) {
-    AccessibilityDelegateUtil.setDelegate(view);
+    ReactAccessibilityDelegate.setDelegate(view);
   }
 
   @Override
   protected void onAfterUpdateTransaction(@Nonnull T view) {
     super.onAfterUpdateTransaction(view);
     updateViewAccessibility(view);
+  }
+
+  @Override
+  public @Nullable Map<String, Object> getExportedCustomDirectEventTypeConstants() {
+    return MapBuilder.<String, Object>builder()
+          .put("performAction", MapBuilder.of("registrationName", "onAccessibilityAction"))
+          .build();
   }
 }
